Big work on overall stuff and things - working compute shaders
I honestly left it so long to create this commit that I forgot what I've done.
main point of this one is working compute shaders and progress on graphics pipelines.

+++ include/WilloRHI/Resources.hpp
@@ -6,10 +6,10 @@
 
 namespace WilloRHI
 {
-    typedef uint64_t BufferId;
-    typedef uint64_t ImageId;
-    typedef uint64_t ImageViewId;
-    typedef uint64_t SamplerId;
+    typedef uint32_t BufferId;
+    typedef uint32_t ImageId;
+    typedef uint32_t ImageViewId;
+    typedef uint32_t SamplerId;
 
     constexpr float LOD_CLAMP_NONE  = 1000.0F;
